+++ src/plugin.js
@@ -62,14 +62,20 @@ function fastifyCloudEvents (fastify, options, next) {
    * @param {object} options optional serialization attributes:
    *        encoder (function, no default) a function that takes data and returns encoded data,
    *        encodedData (string, no default) already encoded data (but consistency with the contentType is not checked),
+   *        onlyValid (boolean, default false) to serialize only if it's a valid instance,
    * @return {string} the serialized event, as a string
    * @throws {Error} if event is undefined or null, or an option is undefined/null/wrong
+   * @throws {Error} if onlyValid is true, and the given event is not a valid CloudEvent instance
    */
-  function serialize (event, { encoder, encodedData } = {}) {
+  function serialize (event, { encoder, encodedData, onlyValid = false } = {}) {
     ensureIsObject(event, 'event')
 
     if (event.contentType === 'application/json') {
-      return stringify(event)
+      if ((onlyValid === false) || (onlyValid === true && CloudEvent.isValidEvent(event) === true)) {
+        return stringify(event)
+      } else {
+        throw new Error(`Unable to serialize a not valid CloudDvent.`)
+      }
     }
     // else
     if (encoder !== undefined && encoder !== null) {
@@ -86,7 +92,12 @@ function fastifyCloudEvents (fastify, options, next) {
     if (typeof encodedData !== 'string') {
       throw new Error(`Missing or wrong encoded data: '${encodedData}' for the given content type: '${event.contentType}'.`)
     }
-    return stringify({ ...event, data: encodedData })
+    const newEvent = { ...event, data: encodedData }
+    if ((onlyValid === false) || (onlyValid === true && CloudEvent.isValidEvent(newEvent) === true)) {
+      return stringify(newEvent)
+    } else {
+      throw new Error(`Unable to serialize a not valid CloudDvent.`)
+    }
   }
 
   /**